requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-jinja2==1.6
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
black==23.9.1
certifi==2025.10.5
cfgv==3.4.0
click==8.3.0
distlib==0.4.0
exceptiongroup==1.3.0
filelock==3.20.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==5.12.0
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
mypy==1.6.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pandas-stubs==2.3.2.250926
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
polars==1.35.2
pre-commit==3.4.0
propcache==0.4.1
pyarrow==17.0.0
Pygments==2.19.2
pytest==7.3.1
python-dotenv==1.0.1
PyYAML==6.0.3
rich==14.2.0
ruff==0.12.1
sniffio==1.3.1
textual==6.5.0
textual-dev==1.8.0
textual-serve==1.1.3
tomli==2.3.0
torch==2.9.1
typing_extensions==4.15.0
uc-micro-py==1.0.3
virtualenv==20.35.4
yarl==1.22.0
